pyproject.toml
[project]
name = "pos-todo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.11.0",
    "python-escpos>=3.1",
]
